Remove files, readd some in
