--- Shared/Extensions/EY+Color.swift
@@ -1,259 +1,0 @@
-//
-//  EY+Color.swift
-//  EngineYard
-//
-//  Created by Amarjit on 26/12/2017.
-//  Copyright © 2017 Amarjit. All rights reserved.
-//
-//  @reference: https://github.com/0x73/SDevFlatColors/blob/master/Classes/SDevFlatColors.swift
-
-import Foundation
-import UIKit
-
-public extension UInt {
-    func convertToUIColor() -> UIColor {
-        return UIColor(
-            red: CGFloat((self & 0xFF0000) >> 16) / 255.0,
-            green: CGFloat((self & 0x00FF00) >> 8) / 255.0,
-            blue: CGFloat(self & 0x0000FF) / 255.0,
-            alpha: CGFloat(1.0)
-        )
-    }
-}
-
-public enum FlatColors {
-    case Turquoise,
-    Emerald,
-    PeterRiver,
-    Amethyst,
-    WetAsphalty,
-    GreenSea,
-    Nephritis,
-    BelizeHole,
-    Wisteria,
-    MidnightBlue,
-    SunFlower,
-    Carrot,
-    Alizarin,
-    Clouds,
-    Concrete,
-    Orange,
-    Pumpkin,
-    Pomegranate,
-    Silver,
-    Asbestos,
-    ChestnuteRose,
-    Thunderbird,
-    OldBrick,
-    Flamingo,
-    Valencia,
-    TallPoppy,
-    Monza,
-    Cinnabar,
-    Razzmatazz,
-    SunsetOrange,
-    WaxFlower,
-    Cabaret,
-    NewYorkPink,
-    RadicalRed,
-    Sunglo,
-    Snuff,
-    Rebeccapurple,
-    HoneyFlower,
-    Wistful,
-    Plum,
-    Seance,
-    MediumPurple,
-    LightWisteria,
-    Studio,
-    SanMarino,
-    AliceBlue,
-    RoyalBlue,
-    PictonBlue,
-    Spray,
-    Shakespeare,
-    HummingBird,
-    Curious,
-    Madison,
-    DodgerBlue,
-    Ming,
-    EbonyClay,
-    Malibu,
-    SummerySky,
-    Chambray,
-    PickledBluewood,
-    Hoki,
-    JellyBean,
-    JacksonsPurple,
-    JordyBlue,
-    SteelBlue,
-    FountainBlue,
-    MediumTurquoise,
-    AquaIsland,
-    Gossip,
-    DarkSeaGreen,
-    Eucalyptus,
-    CaribbeanGreen,
-    SilverTree,
-    Downy,
-    MountainMeadow,
-    LightSeaGreen,
-    MediumAquamarine,
-    Madang,
-    Riptide,
-    Shamrock,
-    Niagara,
-    GreenHaze,
-    OceanGreen,
-    NiagaraOne,
-    Jade,
-    Salem,
-    Observatory,
-    JungleGreen,
-    CapeHoney,
-    California,
-    FireBush,
-    TahitiGold,
-    Casablanca,
-    Crusta,
-    SeaBuckthorn,
-    LightningYellow,
-    BurntOrange,
-    ButterCup,
-    Ecstasy,
-    Sandstorm,
-    Jaffa,
-    Zest,
-    WhiteSmoke,
-    Lynch,
-    Pumice,
-    Gallery,
-    SilverSand,
-    Porcelain,
-    Cascade,
-    Iron,
-    Edward,
-    Cararra
-}
-
-var colors: Dictionary<FlatColors, UInt> = [
-    FlatColors.Turquoise: UInt(0x1abc9c),
-    FlatColors.Emerald: UInt(0x2ecc71),
-    FlatColors.PeterRiver: UInt(0x3498db),
-    FlatColors.Amethyst: UInt(0x9b59b6),
-    FlatColors.WetAsphalty: UInt(0x34495e),
-    FlatColors.GreenSea: UInt(0x16a085),
-    FlatColors.Nephritis: UInt(0x27ae60),
-    FlatColors.BelizeHole: UInt(0x2980b9),
-    FlatColors.Wisteria: UInt(0x8e44ad),
-    FlatColors.MidnightBlue: UInt(0x2c3e50),
-    FlatColors.SunFlower: UInt(0xf1c40f),
-    FlatColors.Carrot: UInt(0xe67e22),
-    FlatColors.Alizarin: UInt(0xe74c3c),
-    FlatColors.Clouds: UInt(0xecf0f1),
-    FlatColors.Concrete: UInt(0x95a5a6),
-    FlatColors.Orange: UInt(0xf39c12),
-    FlatColors.Pumpkin: UInt(0xd35400),
-    FlatColors.Pomegranate: UInt(0xc0392b),
-    FlatColors.Silver: UInt(0xbdc3c7),
-    FlatColors.Asbestos: UInt(0x7f8c8d),
-    FlatColors.ChestnuteRose: UInt(0xd24d57),
-    FlatColors.Thunderbird: UInt(0xD91E18),
-    FlatColors.OldBrick: UInt(0x96281B),
-    FlatColors.Flamingo: UInt(0xEF4836),
-    FlatColors.Valencia: UInt(0xD64541),
-    FlatColors.TallPoppy: UInt(0xC0392B),
-    FlatColors.Monza: UInt(0xCF000F),
-    FlatColors.Cinnabar: UInt(0xE74C3C),
-    FlatColors.Razzmatazz: UInt(0xDB0A5B),
-    FlatColors.SunsetOrange: UInt(0xF64747),
-    FlatColors.WaxFlower: UInt(0xF1A9A0),
-    FlatColors.Cabaret: UInt(0xD2527F),
-    FlatColors.NewYorkPink: UInt(0xE08283),
-    FlatColors.RadicalRed: UInt(0xF62459),
-    FlatColors.Sunglo: UInt(0xE26A6A),
-    FlatColors.Snuff: UInt(0xDCC6E0),
-    FlatColors.Rebeccapurple: UInt(0x663399),
-    FlatColors.HoneyFlower: UInt(0x674172),
-    FlatColors.Wistful: UInt(0xAEA8D3),
-    FlatColors.Plum: UInt(0x913D88),
-    FlatColors.Seance: UInt(0x9A12B3),
-    FlatColors.MediumPurple: UInt(0xBF55EC),
-    FlatColors.LightWisteria: UInt(0xBE90D4),
-    FlatColors.Studio: UInt(0x8E44AD),
-    FlatColors.SanMarino: UInt(0x446CB3),
-    FlatColors.AliceBlue: UInt(0xE4F1FE),
-    FlatColors.RoyalBlue: UInt(0x4183D7),
-    FlatColors.PictonBlue: UInt(0x22A7F0),
-    FlatColors.Spray: UInt(0x81CFE0),
-    FlatColors.Shakespeare: UInt(0x52B3D9),
-    FlatColors.HummingBird: UInt(0xC5EFF7),
-    FlatColors.Curious: UInt(0x3498DB),
-    FlatColors.Madison: UInt(0x2C3E50),
-    FlatColors.DodgerBlue: UInt(0x19B5FE),
-    FlatColors.Ming: UInt(0x336E7B),
-    FlatColors.EbonyClay: UInt(0x22313F),
-    FlatColors.Malibu: UInt(0x6BB9F0),
-    FlatColors.SummerySky: UInt(0x1E8BC3),
-    FlatColors.Chambray: UInt(0x3A539B),
-    FlatColors.PickledBluewood: UInt(0x34495E),
-    FlatColors.Hoki: UInt(0x67809F),
-    FlatColors.JellyBean: UInt(0x2574A9),
-    FlatColors.JacksonsPurple: UInt(0x1F3A93),
-    FlatColors.JordyBlue: UInt(0x89C4F4),
-    FlatColors.SteelBlue: UInt(0x4B77BE),
-    FlatColors.FountainBlue: UInt(0x5C97BF),
-    FlatColors.MediumTurquoise: UInt(0x4ECDC4),
-    FlatColors.AquaIsland: UInt(0xA2DED0),
-    FlatColors.Gossip: UInt(0x87D37C),
-    FlatColors.DarkSeaGreen: UInt(0x90C695),
-    FlatColors.Eucalyptus: UInt(0x26A65B),
-    FlatColors.CaribbeanGreen: UInt(0x03C9A9),
-    FlatColors.SilverTree: UInt(0x68C3A3),
-    FlatColors.Downy: UInt(0x65C6BB),
-    FlatColors.MountainMeadow: UInt(0x1BBC9B),
-    FlatColors.LightSeaGreen: UInt(0x1BA39C),
-    FlatColors.MediumAquamarine: UInt(0x66CC99),
-    FlatColors.Madang: UInt(0xC8F7C5),
-    FlatColors.Riptide: UInt(0x86E2D5),
-    FlatColors.Shamrock: UInt(0x2ECC71),
-    FlatColors.Niagara: UInt(0x16A085),
-    FlatColors.GreenHaze: UInt(0x019875),
-    FlatColors.OceanGreen: UInt(0x4DAF7C),
-    FlatColors.NiagaraOne: UInt(0x2ABB9B),
-    FlatColors.Jade: UInt(0x00B16A),
-    FlatColors.Salem: UInt(0x1E824C),
-    FlatColors.Observatory: UInt(0x049372),
-    FlatColors.JungleGreen: UInt(0x26C281),
-    FlatColors.CapeHoney: UInt(0xFDE3A7),
-    FlatColors.California: UInt(0xF89406),
-    FlatColors.FireBush: UInt(0xEB9532),
-    FlatColors.TahitiGold: UInt(0xE87E04),
-    FlatColors.Casablanca: UInt(0xF4B350),
-    FlatColors.Crusta: UInt(0xF2784B),
-    FlatColors.SeaBuckthorn: UInt(0xEB974E),
-    FlatColors.LightningYellow: UInt(0xF5AB35),
-    FlatColors.BurntOrange: UInt(0xD35400),
-    FlatColors.ButterCup: UInt(0xF39C12),
-    FlatColors.Ecstasy: UInt(0xF9690E),
-    FlatColors.Sandstorm: UInt(0xF9BF3B),
-    FlatColors.Jaffa: UInt(0xF27935),
-    FlatColors.Zest: UInt(0xE67E22),
-    FlatColors.WhiteSmoke: UInt(0xECECEC),
-    FlatColors.Lynch: UInt(0x6C7A89),
-    FlatColors.Pumice: UInt(0xD2D7D3),
-    FlatColors.Gallery: UInt(0xEEEEEE),
-    FlatColors.SilverSand: UInt(0xBDC3C7),
-    FlatColors.Porcelain: UInt(0xECF0F1),
-    FlatColors.Cascade: UInt(0x95A5A6),
-    FlatColors.Iron: UInt(0xDADFE1),
-    FlatColors.Edward: UInt(0xABB7B7),
-    FlatColors.Cararra: UInt(0xF2F1EF)
-];
-
-public extension UIColor {
-    class func flat(color: FlatColors) -> UIColor {
-        let selected: UInt = UInt(colors[color]!)
-        return selected.convertToUIColor()
-    }
-}
