--- Shared/Models/GameModel.swift
@@ -1,11 +1,0 @@
-//
-//  GameModel.swift
-//  EngineYard
-//
-//  Created by Amarjit on 26/01/2018.
-//  Copyright © 2018 Amarjit. All rights reserved.
-//
-
-import Foundation
-import GameplayKit
-
