--- Shared/Models/Player.swift
@@ -1,14 +1,0 @@
-//
-//  Player.swift
-//  EngineYard
-//
-//  Created by Amarjit on 16/01/2018.
-//  Copyright © 2018 Amarjit. All rights reserved.
-//
-
-import Foundation
-
-class Player
-{
-    var tableau: Tableau?
-}
