--- Shared/Models/PlayerModel.swift
@@ -1,10 +1,0 @@
-//
-//  PlayerModel.swift
-//  EngineYard
-//
-//  Created by Amarjit on 26/01/2018.
-//  Copyright © 2018 Amarjit. All rights reserved.
-//
-
-import Foundation
-import GameplayKit
